Move 2 classes into their own file, rename them.
Before, these classes were part of their parent classe's file.
Remove the AGA_ prefix from their name, as they're part of a namespace.
And add basic documentation to their new file.
@todo: PHP Docblocks of these classes.
